+++ rngay-service-authority/src/main/java/com/rngay/service_authority/service/impl/RoleMenuServiceImpl.java
@@ -27,7 +27,7 @@ public class RoleMenuServiceImpl extends ServiceImpl<RoleMenuDao, RoleMenuDTO> i
     private MenuService menuService;
 
     @Override
-    public List<MenuDTO> load(Integer orgId) {
+    public List<MenuDTO> load(Long orgId) {
         return menuList(menuService.loadMenuByOrgId(orgId));
     }
